+++ lib/src/languages/java/java11/java11_generator.dart
@@ -31,6 +31,15 @@ class ApolloCodeGeneratorJava11 extends ApolloCodeGenerator {
     var blockCode = generateASTBlock(f, indent, null, false);
 
     s.write(indent);
+
+    if (f.modifiers.isStatic) {
+      s.write('static ');
+    }
+
+    if (f.modifiers.isFinal) {
+      s.write('final ');
+    }
+
     s.write(typeCode);
     s.write(' ');
     s.write(f.name);